Try making the "to" field at a potentially different index
This isn't good enough. I think there is a bug in that we don't store
enough information in the two field: we need both inner patches with the
join so we can combine them together. In the non-dependent case, nothing
stops us from making this misake, but in the dependent case we would
need to use a `Category` instance on the inner move patches. We plainly
aren't combining two inner patches in the double move case, and the
types let us know.